+++ tests/Integration/ImageManipulatorTest.php
@@ -2,7 +2,6 @@
 
 namespace Plank\Mediable\Tests\Integration;
 
-use _PHPStan_5f1729e44\Nette\Utils\Paginator;
 use Illuminate\Filesystem\FilesystemManager;
 use Intervention\Image\Drivers\Gd\Driver as GdDriver;
 use Intervention\Image\Image;
@@ -10,8 +9,11 @@ use Intervention\Image\ImageManager;
 use Plank\Mediable\Exceptions\ImageManipulationException;
 use Plank\Mediable\ImageManipulation;
 use Plank\Mediable\ImageManipulator;
+use Plank\Mediable\ImageOptimizer;
 use Plank\Mediable\Media;
 use Plank\Mediable\Tests\TestCase;
+use Spatie\ImageOptimizer\Optimizers\Optipng;
+use Spatie\ImageOptimizer\Optimizers\Pngquant;
 
 class ImageManipulatorTest extends TestCase
 {
@@ -697,19 +699,69 @@ class ImageManipulatorTest extends TestCase
         $this->assertSame($expectedVisibility, $result->isVisible());
     }
 
+    public function test_it_can_optimize_images_after_manipulation(): void
+    {
+        if (shell_exec('command -v pngquant') === null) {
+            $this->markTestSkipped('pngquant is not installed.');
+        }
+        if (shell_exec('command -v optipng') === null) {
+            $this->markTestSkipped('optipng is not installed.');
+        }
+
+        $this->useFilesystem('tmp');
+        $this->useDatabase();
+
+        $media = $this->makeMedia(
+            [
+                'disk' => 'tmp',
+                'directory' => 'foo',
+                'filename' => 'bar',
+                'extension' => 'png',
+                'mime_type' => 'image/png',
+                'aggregate_type' => 'image'
+            ]
+        );
+        $this->seedFileForMedia($media, $this->sampleFile());
+        $originalSize = filesize($this->sampleFilePath());
+
+        $manipulation = ImageManipulation::make(function (Image $image) {})
+            ->outputPngFormat()
+            ->optimize([
+                Pngquant::class => [
+                    '--quality=85',
+                    '--force',
+                    '--skip-if-larger'
+                ],
+                Optipng::class => [
+                    '-i0',
+                    '-o2',
+                    '-quiet',
+                ]
+            ]);
+
+        $imageManipulator = $this->getManipulator();
+        $imageManipulator->defineVariant('test', $manipulation);
+        $result = $imageManipulator->createImageVariant($media, 'test');
+
+        $this->assertTrue($result->exists);
+        $this->assertLessThan($originalSize, $result->size);
+    }
+
     public function getManipulator(): ImageManipulator
     {
         if (class_exists(GdDriver::class)) {
             // Intervention Image >=3.0
             return new ImageManipulator(
                 new ImageManager(new GdDriver()),
-                app(FilesystemManager::class)
+                app(FilesystemManager::class),
+                app(ImageOptimizer::class)
             );
         } else {
             // Intervention Image <3.0
             return new ImageManipulator(
                 new ImageManager(['driver' => 'gd']),
-                app(FilesystemManager::class)
+                app(FilesystemManager::class),
+                app(ImageOptimizer::class)
             );
         }
     }